working with binn objects
