+++ .gitignore
@@ -23,3 +23,4 @@ testing/modules/tuningMod/tuningModule.mod
 testing/modules/tuningMod/tuningModule.mod.c
 userspace/user_dtn
 userspace/myclient
+userspace/mybinncli
