+++ userspace/mybinncli.c
@@ -9,10 +9,10 @@
 #include <bits/stdint-uintn.h>
 #include <bits/types.h>
 #include <ctype.h>
-#include <pthread.h>
-#include <sys/prctl.h>
+//#include <pthread.h>
+//#include <sys/prctl.h>
 #include <signal.h>
-#include <sys/wait.h>
+//#include <sys/wait.h>
 
 //#include <linux/bpf.h>
 #include <arpa/inet.h>
@@ -39,9 +39,6 @@ const char *workflow_names[WORKFLOW_NAMES_MAX] = {
 };
 
 FILE * pHpnClientLogPtr = 0;
-FILE * tunLogPtr = 0;
-int IamClient = 1;
-#include "./mybinn.c"
 binn *myobj;
 const char *phase2str(enum workflow_phases phase)
 {
@@ -49,6 +46,37 @@ const char *phase2str(enum workflow_phases phase)
 		return workflow_names[phase];
 return NULL;
 }
+
+void fMake_Binn_Object(struct PeerMsg *pMsg, binn * obj)
+{
+	binn_object_set_uint32(obj, "msg_no", pMsg->msg_no);
+	binn_object_set_uint32(obj, "seq_no", pMsg->seq_no);
+	binn_object_set_uint32(obj, "value", pMsg->value);
+	binn_object_set_uint32(obj, "hop_latency", pMsg->hop_latency);
+	binn_object_set_uint32(obj, "queue_occupancy", pMsg->queue_occupancy);
+	binn_object_set_uint32(obj, "switch_id", pMsg->switch_id);
+	binn_object_set_str(obj, "timestamp", pMsg->timestamp);
+	binn_object_set_str(obj, "msg", pMsg->msg);
+
+return;
+}
+
+void fRead_Binn_Object(struct PeerMsg *pMsg, binn * obj)
+{
+
+	pMsg->msg_no = binn_object_uint32(obj, "msg_no");
+	pMsg->seq_no = binn_object_uint32(obj, "seq_no");
+	pMsg->value = binn_object_uint32(obj, "value");
+	pMsg->hop_latency = binn_object_uint32(obj, "hop_latency");
+	pMsg->queue_occupancy = binn_object_uint32(obj, "queue_occupancy");
+	pMsg->switch_id = binn_object_uint32(obj, "switch_id");
+	pMsg->ptimes = binn_object_str(obj, "timestamp");
+	pMsg->pm = binn_object_str(obj, "msg");
+
+	return;
+}
+
+
 
 
 int vDebugLevel = 1;
